Fixed a lot more warnings.

+++ java/component/keyple-common/src/main/java/org/eclipse/keyple/common/stringhelper.h
@@ -298,7 +298,7 @@ class StringHelper {
         int percent        = -1;
         while ((percent = input.find('%', percent + 1)) > -1)
         {
-            if (percent + 1 < input.length())
+            if (percent + 1 < (int)input.length())
             {
                 if (input[percent + 1] == '%')
                 {
@@ -308,7 +308,7 @@ class StringHelper {
 
                 int formatEnd = -1;
                 std::string index;
-                for (int i = percent + 1; i < input.length(); i++)
+                for (int i = percent + 1; i < (int)input.length(); i++)
                 {
                     if (input[i] == 's')
                     {
@@ -316,7 +316,7 @@ class StringHelper {
                         formatEnd = i;
                         break;
                     }
-                    else if (input[i] == '$' && i + 1 < input.length() && input[i + 1] == 's')
+                    else if (input[i] == '$' && i + 1 < (int)input.length() && input[i + 1] == 's')
                     {
                         index     = input.substr(percent + 1, i - percent - 1);
                         formatEnd = i + 1;
@@ -346,7 +346,7 @@ class StringHelper {
             }
         }
 
-        if (lastFormatChar + 1 < input.length())
+        if (lastFormatChar + 1 < (int)input.length())
             ss << input.substr(lastFormatChar + 1);
 
         return ss.str();
